fix type conversion

+++ cpu_2d/vreduce.hpp
@@ -88,7 +88,7 @@ void vreduce(function<void(T, long, T *, int)> &reduce,
             reduce(0, ssize, uncompressed_fq, uncompressedsize);
 
 #ifdef _COMPRESSION
-            if (isCompressed(originalsize, psize_from))
+            if (isCompressed(static_cast<size_t>(originalsize), static_cast<size_t>(psize_from)))
             {
                 if (uncompressed_fq != NULL)
                 {
@@ -228,7 +228,7 @@ void vreduce(function<void(T, long, T *, int)> &reduce,
                 reduce(offset, lowerId, uncompressed_fq, uncompressedsize);
 
 #ifdef _COMPRESSION
-                if (isCompressed(originalsize, psizeFrom))
+                if (isCompressed(static_cast<size_t>(originalsize), static_cast<size_t>(psizeFrom)))
                 {
                     if (uncompressed_fq != NULL)
                     {
@@ -301,7 +301,7 @@ void vreduce(function<void(T, long, T *, int)> &reduce,
                 reduce(offset + lowerId, upperId, uncompressed_fq, uncompressedsize);
 
 #ifdef _COMPRESSION
-                if (isCompressed(originalsize, psizeFrom))
+                if (isCompressed(static_cast<size_t>(originalsize), static_cast<size_t>(psizeFrom)))
                 {
                     if (uncompressed_fq != NULL)
                     {
